# docs/api/openapi.yaml
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HumanAction": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "default": 0,
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "rationale": {
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          }
        },
        "required": [
          "rationale"
        ],
        "title": "HumanAction",
        "type": "object"
      },
      "PortfolioRunRequest": {
        "additionalProperties": false,
        "properties": {
          "data_cutoff": {
            "format": "date-time",
            "title": "Data Cutoff",
            "type": "string"
          },
          "decision_time": {
            "format": "date-time",
            "title": "Decision Time",
            "type": "string"
          }
        },
        "required": [
          "decision_time",
          "data_cutoff"
        ],
        "title": "PortfolioRunRequest",
        "type": "object"
      },
...
      "RunResponse": {
        "properties": {
          "data_cutoff": {
            "format": "date-time",
            "title": "Data Cutoff",
            "type": "string"
          },
          "decision_time": {
            "format": "date-time",
            "title": "Decision Time",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "run_type": {
            "enum": [
              "RESEARCH",
              "PORTFOLIO"
            ],
            "title": "Run Type",
            "type": "string"
          },
          "status": {
            "enum": [
              "QUEUED",
              "RUNNING",
              "COMPLETED",
              "FAILED",
              "CANCELLED"
            ],
            "title": "Status",
            "type": "string"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "required": [
          "run_id",
          "run_type",
          "status",
          "symbol",
          "decision_time",
          "data_cutoff"
        ],
        "title": "RunResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AI Stock Research Platform",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/alerts/{alert_id}/acknowledge": {
      "post": {
        "operationId": "acknowledge_alert_api_v1_alerts__alert_id__acknowledge_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Acknowledge Alert Api V1 Alerts  Alert Id  Acknowledge Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Acknowledge Alert"
      }
    },
...
    "/api/v1/evals/runs/{eval_run_id}": {
      "get": {
        "operationId": "get_eval_run_api_v1_evals_runs__eval_run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "eval_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Eval Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Eval Run Api V1 Evals Runs  Eval Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Eval Run"
      }
    },
...
    "/api/v1/policies/{policy_id}/activate": {
      "post": {
        "operationId": "activate_policy_api_v1_policies__policy_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Policy Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Activate Policy Api V1 Policies  Policy Id  Activate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Policy"
      }
    },
    "/api/v1/policies/{policy_id}/rollback": {
      "post": {
        "operationId": "rollback_policy_api_v1_policies__policy_id__rollback_post",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Policy Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Rollback Policy Api V1 Policies  Policy Id  Rollback Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rollback Policy"
      }
    },
...
    "/api/v1/portfolio/rebalance-runs": {
      "post": {
        "operationId": "create_portfolio_run_api_v1_portfolio_rebalance_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Portfolio Run"
      }
    },
...
    "/api/v1/research-runs": {
      "post": {
        "operationId": "create_research_run_api_v1_research_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Research Run"
      }
    },
    "/api/v1/research-runs/{run_id}": {
      "get": {
        "operationId": "get_research_run_api_v1_research_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Research Run"
      }
    },
    "/api/v1/research-runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_research_run_api_v1_research_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Research Run"
      }
    },
    "/api/v1/research-runs/{run_id}/report": {
      "get": {
        "operationId": "get_research_report_api_v1_research_runs__run_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Research Report Api V1 Research Runs  Run Id  Report Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Research Report"
      }
    },
    "/api/v1/stocks/{symbol}/research": {
      "get": {
        "operationId": "get_stock_research_api_v1_stocks__symbol__research_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Stock Research Api V1 Stocks  Symbol  Research Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Stock Research"
      }
    },
    "/api/v1/watchlist": {
      "get": {
        "operationId": "list_watchlist_api_v1_watchlist_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Watchlist Api V1 Watchlist Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Watchlist"
      },
      "post": {
        "operationId": "add_watchlist_api_v1_watchlist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Watchlist Api V1 Watchlist Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Watchlist"
      }
    },
    "/api/v1/watchlist/{symbol}": {
      "delete": {
        "operationId": "delete_watchlist_api_v1_watchlist__symbol__delete",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Watchlist"
      },
      "patch": {
        "operationId": "patch_watchlist_api_v1_watchlist__symbol__patch",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Patch Watchlist Api V1 Watchlist  Symbol  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Watchlist"
      }
    },
...
  }
}